1544 - Moved job role cleaning utils into their own script. moved tests and updated imports accordingly.
