+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/tests/fargate/utils/test_clean_utils.py
@@ -1,0 +1,64 @@
+import unittest
+
+import polars as pl
+import polars.testing as pl_testing
+
+import projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.utils.clean_utils as job
+from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
+from utils.column_values.categorical_column_values import EstimateFilledPostsSource
+
+
+class NullifyJobRoleCountWhenSourceNotAscwds(unittest.TestCase):
+    def setUp(self) -> None:
+        self.test_schema = {
+            IndCQC.ascwds_filled_posts_dedup_clean: pl.Float64,
+            IndCQC.estimate_filled_posts: pl.Float64,
+            IndCQC.estimate_filled_posts_source: pl.String,
+            IndCQC.ascwds_job_role_counts: pl.Int64,
+        }
+        self.input_rows_that_meet_condition = [
+            (10.0, 10.0, EstimateFilledPostsSource.ascwds_pir_merged, 1),
+            (10.0, 10.0, EstimateFilledPostsSource.ascwds_pir_merged, 2),
+        ]
+
+        self.expected_rows_that_meet_condition = self.input_rows_that_meet_condition
+
+    def _create_input_lf(self, extra_rows: list[tuple]) -> pl.LazyFrame:
+        """Set the input LazyFrame up with rows that meet condition + given rows."""
+        return pl.LazyFrame(
+            [*self.input_rows_that_meet_condition, *extra_rows],
+            self.test_schema,
+            orient="row",
+        )
+
+    def _create_expected_lf(self, extra_rows: list[tuple]) -> pl.LazyFrame:
+        """Set the expected LazyFrame up with rows that meet condition + given rows."""
+        return pl.LazyFrame(
+            [*self.expected_rows_that_meet_condition, *extra_rows],
+            self.test_schema,
+            orient="row",
+        )
+
+    def test_nullifies_when_source_not_ascwds(self):
+        input_lf = self._create_input_lf(
+            # Row with non-ASCWDS source
+            [(10.0, 10.0, EstimateFilledPostsSource.care_home_model, 2)],
+        )
+        expected_lf = self._create_expected_lf(
+            # Row with non-ASCWDS source (counts nullified)
+            [(10.0, 10.0, EstimateFilledPostsSource.care_home_model, None)],
+        )
+        returned_lf = job.nullify_job_role_count_when_source_not_ascwds(input_lf)
+        pl_testing.assert_frame_equal(returned_lf, expected_lf)
+
+    def test_nullifies_when_estimate_doesnt_match_ascwds(self):
+        input_lf = self._create_input_lf(
+            # Row that doesn't match estimate filled posts.
+            [(None, 20.0, EstimateFilledPostsSource.ascwds_pir_merged, 1)],
+        )
+        expected_lf = self._create_expected_lf(
+            # Row that doesn't match estimate filled posts (counts nullified)
+            [(None, 20.0, EstimateFilledPostsSource.ascwds_pir_merged, None)],
+        )
+        returned_lf = job.nullify_job_role_count_when_source_not_ascwds(input_lf)
+        pl_testing.assert_frame_equal(returned_lf, expected_lf)
